+++ app/Http/Controllers/Admin/TopInstituteController.php
@@ -7,6 +7,7 @@ use Illuminate\Http\Request;
 use App\Http\Controllers\Controller;
 use App\Http\Requests\Admin\TopInstituteRequest;
 use App\Models\Institute;
+use Illuminate\Support\Facades\Gate;
 use Inertia\Inertia;
 
 class TopInstituteController extends Controller
@@ -16,6 +17,7 @@ class TopInstituteController extends Controller
      */
     public function index()
     {
+        Gate::authorize('viewAny', TopInstitute::class);
         $topInstitute = TopInstitute::with('institute')->paginate(50);
         return Inertia::render('Admin/Institute/Top/List', compact('topInstitute'));
     }
@@ -25,6 +27,7 @@ class TopInstituteController extends Controller
      */
     public function create()
     {
+        Gate::authorize('create', TopInstitute::class);
         $institute = Institute::all();
         return Inertia::render('Admin/Institute/Top/Form', compact('institute'));
     }
@@ -34,6 +37,7 @@ class TopInstituteController extends Controller
      */
     public function store(TopInstituteRequest $request)
     {
+        Gate::authorize('create', TopInstitute::class);
         $validated = $request->validated();
         TopInstitute::create($validated);
         return redirect()->route('admin.top_institute.index')->with('success', 'Created succesfully!');
@@ -52,6 +56,7 @@ class TopInstituteController extends Controller
      */
     public function edit(TopInstitute $topInstitute)
     {
+        Gate::authorize('update', $topInstitute);
         $institute = Institute::all();
         return Inertia::render('Admin/Institute/Top/Form', compact('institute', 'topInstitute'));
     }
@@ -61,6 +66,7 @@ class TopInstituteController extends Controller
      */
     public function update(TopInstituteRequest $request, TopInstitute $topInstitute)
     {
+        Gate::authorize('update', $topInstitute);
         $validated = $request->validated();
         $topInstitute->update($validated);
         return redirect()->route('admin.top_institute.index')->with('success', 'Updated successfully!');
@@ -71,6 +77,7 @@ class TopInstituteController extends Controller
      */
     public function destroy(TopInstitute $topInstitute)
     {
+        Gate::authorize('delete', $topInstitute);
         $topInstitute->delete();
         return redirect()->route('admin.top_institute.index')->with('success', 'Deleted succesfully!');
     }
